refactor: create user form
