+++ web/src/components/CreateUserForm/CreateUserForm.stories.tsx
@@ -1,0 +1,25 @@
+// When you've added props to your component,
+// pass Storybook's `args` through this story to control it from the addons panel:
+//
+// ```tsx
+// import type { ComponentStory } from '@storybook/react'
+//
+// export const generated: ComponentStory<typeof CreateUserForm> = (args) => {
+//   return <CreateUserForm {...args} />
+// }
+// ```
+//
+// See https://storybook.js.org/docs/react/writing-stories/args.
+
+import type { ComponentMeta } from '@storybook/react'
+
+import CreateUserForm from './CreateUserForm'
+
+export const generated = () => {
+  return <CreateUserForm />
+}
+
+export default {
+  title: 'Components/CreateUserForm',
+  component: CreateUserForm,
+} as ComponentMeta<typeof CreateUserForm>
